requirements.txt
Flask==3.1.0
gunicorn==23.0.0
python-crontab==3.2.0
python-dotenv==1.0.1
